+++ src/games/chlcc/tipssystem.h
@@ -9,16 +9,16 @@ using namespace Impacto::TipsSystem;
 
 class TipsSystem : public TipsSystemBase {
  public:
-  TipsSystem(int maxTipsCount) : TipsSystemBase(maxTipsCount) {};
+  TipsSystem(size_t maxTipsCount) : TipsSystemBase(maxTipsCount) {};
 
   void DataInit(uint32_t scriptBufferId, uint32_t tipsDataAdr,
                 uint32_t tipsDataSize);
   void UpdateTipRecords();
-  void SetTipLockedState(int id, bool state);
-  void SetTipUnreadState(int id, bool state);
-  void SetTipNewState(int id, bool state);
+  void SetTipLockedState(size_t id, bool state);
+  void SetTipUnreadState(size_t id, bool state);
+  void SetTipNewState(size_t id, bool state);
 
-  bool GetTipLockedState(int id);
+  bool GetTipLockedState(size_t id);
 };
 
 }  // namespace CHLCC
